fix: 薪资模块金额字段类型错误导致小数被截断
将员工薪资和社保相关字段从 Long 改为 BigDecimal，解决前端
传入小数（如 1.5）被截断为整数（1）的问题。

修复范围：
- HrmEmployeeSalary (Domain/Bo/Vo): 9 个金额字段
- HrmEmployeeSocialInsurance (Domain/Bo/Vo): 16 个金额字段

涉及字段：
- 薪资：基薪、绩效工资、证书费、岗位工资、加班工资、各类补助等
- 社保：养老/医疗/失业/工伤/生育保险、住房公积金、专项附加扣除等

+++ ruoyi-admin/src/main/java/org/dromara/hrm/money/domain/HrmEmployeeSalary.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import java.io.Serial;
+import java.math.BigDecimal;
 
 /**
  * 员工薪资对象 hrm_employee_salary
@@ -61,47 +62,47 @@ public class HrmEmployeeSalary extends TenantEntity {
     /**
      * 基薪
      */
-    private Long basicSalary;
+    private BigDecimal basicSalary;
 
     /**
      * 月绩效工资
      */
-    private Long monthlyPerformanceSalary;
+    private BigDecimal monthlyPerformanceSalary;
 
     /**
      * 证书费
      */
-    private Long certificateFee;
+    private BigDecimal certificateFee;
 
     /**
      * 岗位工资
      */
-    private Long postSalary;
+    private BigDecimal postSalary;
 
     /**
      * 加班工资
      */
-    private Long overtimeSalary;
+    private BigDecimal overtimeSalary;
 
     /**
      * 通讯补助
      */
-    private Long communicationSubsidy;
+    private BigDecimal communicationSubsidy;
 
     /**
      * 交通补助
      */
-    private Long trafficSubsidy;
+    private BigDecimal trafficSubsidy;
 
     /**
      * 餐补/误餐
      */
-    private Long mealSubsidy;
+    private BigDecimal mealSubsidy;
 
     /**
      * 其他加款
      */
-    private Long otherAdditional;
+    private BigDecimal otherAdditional;
 
     /**
      * 薪资月份（格式：YYYYMM）
