+++ ruoyi-admin/src/main/java/org/dromara/hrm/money/domain/bo/HrmEmployeeSalaryBo.java
@@ -8,6 +8,7 @@ import io.github.linpeilie.annotations.AutoMapper;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 /**
  * 员工薪资业务对象 hrm_employee_salary
@@ -56,47 +57,47 @@ public class HrmEmployeeSalaryBo extends BaseEntity {
     /**
      * 基薪
      */
-    private Long basicSalary;
+    private BigDecimal basicSalary;
 
     /**
      * 月绩效工资
      */
-    private Long monthlyPerformanceSalary;
+    private BigDecimal monthlyPerformanceSalary;
 
     /**
      * 证书费
      */
-    private Long certificateFee;
+    private BigDecimal certificateFee;
 
     /**
      * 岗位工资
      */
-    private Long postSalary;
+    private BigDecimal postSalary;
 
     /**
      * 加班工资
      */
-    private Long overtimeSalary;
+    private BigDecimal overtimeSalary;
 
     /**
      * 通讯补助
      */
-    private Long communicationSubsidy;
+    private BigDecimal communicationSubsidy;
 
     /**
      * 交通补助
      */
-    private Long trafficSubsidy;
+    private BigDecimal trafficSubsidy;
 
     /**
      * 餐补/误餐
      */
-    private Long mealSubsidy;
+    private BigDecimal mealSubsidy;
 
     /**
      * 其他加款
      */
-    private Long otherAdditional;
+    private BigDecimal otherAdditional;
 
     /**
      * 薪资月份（格式：YYYYMM）
